# Chore: Remove prettier plugin