--- a/package.json
+++ b/package.json
@@ -26,10 +26,8 @@
     "eslint": "7.28.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-jsdoc": "37.7.0",
-    "eslint-plugin-prettier": "4.0.0",
     "eslint-plugin-react": "7.28.0",
     "eslint-plugin-react-hooks": "4.3.0",
-    "prettier": "2.5.1",
     "typescript": "4.4.4"
   }
 }