--- a/package.json
+++ b/package.json
@@ -2,7 +2,7 @@
   "author": "Grafana Labs",
   "license": "Apache-2.0",
   "name": "@grafana/eslint-config",
-  "version": "2.5.2",
+  "version": "3.0.0",
   "description": "Grafana's ESLint config",
   "keywords": [
     "grafana",