The prettier plugin is quite slow, prettier recommend running it [separately to linters](https://prettier.io/docs/en/integrating-with-linters.html). From profiling the Grafana linting command we can see it consumes ~73% of time taken to complete:

BEFORE:

```TIMING=1 yarn eslint . --ext .js,.tsx,.ts
Rule                                       | Time (ms) | Relative
:------------------------------------------|----------:|--------:
prettier/prettier                          | 47121.683 |    73.1%
jsdoc/check-alignment                      |  3051.863 |     4.7%
react/no-deprecated                        |  2758.311 |     4.3%
react/display-name                         |  2591.744 |     4.0%
react/no-direct-mutation-state             |  2117.436 |     3.3%
react/no-string-refs                       |  1578.806 |     2.4%
react/require-render-return                |  1235.619 |     1.9%
@typescript-eslint/naming-convention       |  1076.987 |     1.7%
no-redeclare                               |   320.645 |     0.5%
@typescript-eslint/type-annotation-spacing |   239.294 |     0.4%
```
AFTER:
```
Rule                                       | Time (ms) | Relative
:------------------------------------------|----------:|--------:
jsdoc/check-alignment                      |  3698.093 |    18.0%
react/no-deprecated                        |  3270.812 |    15.9%
react/display-name                         |  3054.202 |    14.9%
react/no-direct-mutation-state             |  2677.127 |    13.0%
react/no-string-refs                       |  1943.171 |     9.5%
react/require-render-return                |  1493.245 |     7.3%
@typescript-eslint/naming-convention       |  1104.044 |     5.4%
no-redeclare                               |   384.001 |     1.9%
jest/no-focused-tests                      |   319.223 |     1.6%
@typescript-eslint/type-annotation-spacing |   230.898 |     1.1%
```